Rework SQL formatting using PgQuery#scan
try a completely new approach that should be more resilient. The idea
is to have an Indenter that only adds extra whitespace on top of
deparsed string, guaranteeing there will be no syntax error.

this is a slightly naive but quite simple approach that should deal
with the 99% case pretty well, and deals gracefully with the rest.